cmd/geth: move CoreGeth-specific test to non-conflicting file
Date: 2021-03-19 10:44:55-05:00
